+++ TheMidway/TheMidway/controller/NovoEncontroViewController.swift
@@ -31,8 +31,7 @@ class NovoEncontroViewController: UIViewController, CLLocationManagerDelegate, M
         localLabel.isHidden = true
         
         
-        
-        // Do any additional setup after loading the view.
+        /* MapKit */
         
         // Define o delegate do mapa
         self.mapView.delegate = self
@@ -40,7 +39,10 @@ class NovoEncontroViewController: UIViewController, CLLocationManagerDelegate, M
         // Define o delegate das localizações
         self.locationManager.delegate = self
         self.locationManager.requestWhenInUseAuthorization()
+        
+        
     }
+    
     
     // MARK: Prepare
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
@@ -50,12 +52,12 @@ class NovoEncontroViewController: UIViewController, CLLocationManagerDelegate, M
         }
     }
     
-    // MARK: MApView
-    ///Atributos
-    /// View da controller
     
+    
+    // MARK: - MapView
+
     /// Tipos de lugares que queremos pra busca
-    private let placesTypes: [MKPointOfInterestCategory] = [
+    public let placesTypes: [MKPointOfInterestCategory] = [
         // Parques
         .amusementPark,
         .nationalPark,
@@ -72,10 +74,10 @@ class NovoEncontroViewController: UIViewController, CLLocationManagerDelegate, M
     ]
     
     /// Locais encontrados
-    private var nerbyPlaces: [MapPlace] = []
+    public var nerbyPlaces: [MapPlace] = []
     
     /// Lidando com a localização
-    private lazy var locationManager: CLLocationManager = {
+    public lazy var locationManager: CLLocationManager = {
         var manager = CLLocationManager()
         manager.distanceFilter = 10
         manager.desiredAccuracy = kCLLocationAccuracyBest
@@ -83,7 +85,7 @@ class NovoEncontroViewController: UIViewController, CLLocationManagerDelegate, M
     }()
     
     /// Pontos adicionados como referência
-    private let pointsExample: [CLLocationCoordinate2D] = [
+    public let pointsExample: [CLLocationCoordinate2D] = [
         CLLocationCoordinate2D(latitude: -23.495480, longitude: -46.868080),    // Muza
         CLLocationCoordinate2D(latitude: -23.545580, longitude: -46.651860),    // Feh
         CLLocationCoordinate2D(latitude: -23.523580, longitude: -46.774770),    // Bia
@@ -93,16 +95,17 @@ class NovoEncontroViewController: UIViewController, CLLocationManagerDelegate, M
     ]
     
     /// Coordenaa do ponto médio achado
-    private var midpoint: CLLocationCoordinate2D = CLLocationCoordinate2D(latitude: 0, longitude: 0)
+    public var midpoint: CLLocationCoordinate2D = CLLocationCoordinate2D(latitude: 0, longitude: 0)
     
     /// Tamanho da área que vai ser desenhada pra fazer a busca
-    private let radiusArea: CLLocationDistance = 2500
+    public let radiusArea: CLLocationDistance = 2500
     
     /// Palavras que vão ser buscadas na hora de pesquisar os lugares
-    private let searchWords: [String] = ["bar", "restaurant", "pizza", "shopping", "club", "park", "night", "party"]
+    public let searchWords: [String] = ["bar", "restaurant", "pizza", "shopping", "club", "park", "night", "party"]
         
     /// Pontos encotrados por um endereço como string
-    private var coordFound: [CLLocationCoordinate2D] = []
+    public var coordFound: [CLLocationCoordinate2D] = []
+    
     
     
     
@@ -151,249 +154,34 @@ class NovoEncontroViewController: UIViewController, CLLocationManagerDelegate, M
         if self.nerbyPlaces.isEmpty {
             // Pega o ponto central
             self.midpoint = self.theMidpoint(coordinates: self.pointsExample)
-            
+
             // Adiciona os pontos exemplos
             for coords in self.pointsExample {
                 self.addPointOnMap(pin: self.createPin(name: " ", coordinate: coords))
             }
-            
+
             // Adiciona o ponto central
             self.addPointOnMap(pin: self.createPin(name: " ", coordinate: self.midpoint))
-            
+
             // Cria um cículo
             self.addCircle(location: self.midpoint)
-            
+
             // Define a região que vai ser focada no mapa: o ponto dentral
             self.setViewLocation(place: self.midpoint, radius: self.radiusArea)
         }
-        
+
         // Faz a busca por locasi a partr das palavras chaves.
         for someWord in self.searchWords {
             self.getNerbyPlaces(someWord)
         }
-        
+
         // self.getCoordsByAddress(address: "Rua Nicola Spinelli, 469")
     }
-    
-    
-    /// Tira o teclado
-    @objc func dismissKeyboard() {self.view.endEditing(true)}
-    
-    
-    
-    
-    /* MARK: - Pins */
-    
-    /// Adicionar ponto no mapa
-    private func addPointOnMap(pin: MKPointAnnotation) -> Void {
-        self.mapView.addAnnotation(pin)
-    }
-    
-    /// Cria um pin
-    private func createPin(name: String, coordinate: CLLocationCoordinate2D) -> MKPointAnnotation {
-        let pin = MKPointAnnotation()
-        pin.coordinate = coordinate
-        pin.title = name
-        // pin.subtitle = ""
-        
-        return pin
-    }
-    
-    
-    
-    
-    /* MARK: - Gerenciamento do Mapa */
-    
-    /// Define a região que o mapa vai mostrar
-    private func setViewLocation(place: CLLocationCoordinate2D, radius: CLLocationDistance) -> Void {
-        // let radiusDistance = CLLocationDistance(exactly: radius)!
-        let region = MKCoordinateRegion(center: place, latitudinalMeters: radius, longitudinalMeters: radius)
-        
-        self.mapView.setRegion(region, animated: true)
-    }
-        
-    
-    
-    
-    /* MARK: - Ponto Médio */
-    
-    /// Calculo do ponto médio entre os pontos
-    private func theMidpoint(coordinates: [CLLocationCoordinate2D]) -> CLLocationCoordinate2D {
-        let coordsCount = Double(coordinates.count)
-        
-        guard coordsCount > 1 else {
-            return coordinates.first ?? CLLocationCoordinate2D(latitude: 0, longitude: 0)
-        }
-
-        var x: Double = 0.0
-        var y: Double = 0.0
-        var z: CGFloat = 0.0
-
-        for coordinate in coordinates {
-            let lat: CGFloat = self.toRadian(angle: coordinate.latitude)
-            let lon: CGFloat = self.toRadian(angle: coordinate.longitude)
-            
-            x += cos(lat) * cos(lon)
-            y += cos(lat) * sin(lon)
-            z += sin(lat)
-        }
-                
-        x /= coordsCount
-        y /= coordsCount
-        z /= coordsCount
-
-        let lon = atan2(y, x)
-        let hyp = sqrt(x*x + y*y)
-        let lat = atan2(z, hyp)
-
-        return CLLocationCoordinate2D(latitude: self.toDegree(radian: lat), longitude: self.toDegree(radian: lon))
-    }
-    
-    /* Degrees to Radian */
-    private func toRadian(angle: CLLocationDegrees) -> CGFloat {
-        return CGFloat(angle) / 180.0 * CGFloat(Double.pi)
-    }
-
-    /* Radians to Degrees */
-    private func toDegree(radian: CGFloat) -> CLLocationDegrees {
-        return CLLocationDegrees(radian * CGFloat(180.0 / Double.pi))
-    }
-    
-    
-    
-    
-    /* MARK: - Desenhar */
-    
-    /// Adiciona um cículo
-    private func addCircle(location: CLLocationCoordinate2D) -> Void {
-        let circle = MKCircle(center: location, radius: self.radiusArea as CLLocationDistance)
-        
-        self.mapView.addOverlay(circle)
-    }
-    
-    
-    
-    
-    /* MARK: - Locais */
-    
-    /// Pega os lugares próximos a partir
-    private func getNerbyPlaces(_ mainWord: String) -> Void {
-        // Cria a pesquisa:
-        let request = MKLocalSearch.Request()
-        
-        request.naturalLanguageQuery = mainWord
-        
-        // Coloca os pontos que deseja
-        request.pointOfInterestFilter = MKPointOfInterestFilter(including: self.placesTypes)
-        
-        // Define a regiao da pesquisa (a mesma do mapa)
-        request.region = MKCoordinateRegion(center: self.midpoint, latitudinalMeters: 400, longitudinalMeters: 400)
-                
-        // Realiza a pesquisa
-        let search = MKLocalSearch(request: request)
-        
-        search.start() { response, error in
-            
-            guard let response = response else {
-                // Lidar com o erro aqui!
-                print("Estou no erro.")
-                return
-            }
-            
-            print("Itens encontrados para \(mainWord): \(response.mapItems.count). \n\n")
-                        
-            for item in response.mapItems {
-                if let name = item.name, let location = item.placemark.location, let type = item.pointOfInterestCategory {
-                    // Cria a cordenada
-                    let coords = CLLocationCoordinate2D(
-                        latitude: location.coordinate.latitude,
-                        longitude: location.coordinate.longitude
-                    )
-                    
-                    // Cria o pin
-                    let pin = self.createPin(name: name, coordinate: coords)
-                    
-                    // Guarda as informações
-                    let newItem: MapPlace = MapPlace(
-                        name: name,
-                        coordinates: coords,
-                        address: location,
-                        pin: pin,
-                        type: type
-                    )
-                    
-                    // Adiciona na lista
-                    
-                    // Verifica se está dentor do raio E se já não foi adicionado.
-                    if self.getDistance(place: coords) <= self.radiusArea+500 && !self.findPlace(place: newItem){
-                        self.nerbyPlaces.append(newItem)
-                        
-                        print("Lugar: \(newItem.name)")
-                        print("Site: \(String(describing: item.url))")
-                        print("Tipo: \(String(describing: item.pointOfInterestCategory!))\n")
-                        
-                        // Add no mapa
-                        self.addPointOnMap(pin: pin)
-                    }
-                }
-            }
-            
-            print("\n\n Locais no mapa: \(self.nerbyPlaces.count) \n\n")
-        }
-    }
-    
-    /// Calcula a distancia entre dois pontos em metros
-    private func getDistance(place: CLLocationCoordinate2D) -> Double {
-        let midpoint = CLLocation(latitude: self.midpoint.latitude, longitude: self.midpoint.longitude)
-        let newPlace = CLLocation(latitude: place.latitude, longitude: place.longitude)
-        
-        return midpoint.distance(from: newPlace)
-    }
-    
-    /// Verifica se já foi achado esse lugar
-    private func findPlace(place: MapPlace) -> Bool {
-        for places in self.nerbyPlaces {
-            if place.name == places.name {
-                return true
-            }
-        }
-        return false
-    }
-    
-    
-    /// Pega a cordenada a partir de um endereço
-    private func getCoordsByAddress(address: String) -> Void {
-        let geocoder: CLGeocoder = CLGeocoder()
-        
-        geocoder.geocodeAddressString(address) { placemarks, error in
-            
-            // Lidando com o erro.
-            if let _ = error {
-                return
-            }
-            
-            // Caso tenha algum resultado um resultado
-            if (placemarks?.count ?? 0 > 0) {
-                // Pega o primeiro resultado
-                let topResult: CLPlacemark = (placemarks?[0])!
-                let placemark: MKPlacemark = MKPlacemark(placemark: topResult)
-                
-                let point = CLLocationCoordinate2D(latitude: (placemark.location?.coordinate.latitude)!,
-                                                   longitude: (placemark.location?.coordinate.longitude)!)
-                self.coordFound.append(point)
-                
-                for c in self.coordFound{
-                    let pin = self.createPin(name: "", coordinate: c)
-                    self.addPointOnMap(pin: pin)
-                }
-            } else {
-                print("Não foi achado nenhum endereço.")
-            }
-        }
-    }
 }
-// MARK: Table View
-extension NovoEncontroViewController: UITableViewDelegate{
+
+
+// MARK: - Table View
+extension NovoEncontroViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         ///clique na celula
         tableView.deselectRow(at: indexPath, animated: true)
@@ -401,7 +189,9 @@ extension NovoEncontroViewController: UITableViewDelegate{
     }
 }
 
-extension NovoEncontroViewController: UITableViewDataSource{
+
+
+extension NovoEncontroViewController: UITableViewDataSource {
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         var cellBase = UITableViewCell()
         if indexPath.row == 0{
@@ -427,26 +217,31 @@ extension NovoEncontroViewController: UITableViewDataSource{
     }
 }
 
-// MARK: Collection View
-extension NovoEncontroViewController: UICollectionViewDelegate{
+
+
+// MARK: - Collection View
+extension NovoEncontroViewController: UICollectionViewDelegate {
     
 }
 
-extension NovoEncontroViewController:UICollectionViewDataSource{
+extension NovoEncontroViewController:UICollectionViewDataSource {
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         return 10
     }
     
     func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "novoEncontroCollection", for: indexPath) as! NovoEncontroCollectionViewCell
-        if nerbyPlaces.count != 0 {
-            cell.stylize(nearbyPlace: nerbyPlaces[indexPath.row])
+        if self.nerbyPlaces.count != 0 {
+            cell.stylize(nearbyPlace: self.nerbyPlaces[indexPath.row])
         }
         return cell
     }
     
     
 }
+
+
+
 
 // MARK: Delegate
 
@@ -470,5 +265,9 @@ extension NovoEncontroViewController: QuemVaiViewControllerDelegate{
         //self.mapView.getButton().addTarget(self, action: #selector(self.buttonAction), for: .touchDown)
     }
 }
+
+
+
+
