+++ TheMidway/TheMidway/model/Pessoa.swift
@@ -1,0 +1,13 @@
+//
+//  Pessoa.swift
+//  TheMidway
+//
+//  Created by Gui Reis on 30/11/21.
+//
+
+/// Informações sobre uma pessoa
+struct Pessoa {
+    var nome: String
+    var endereco: String
+    var icone: String
+}
